requirements.txt
keras==2.9.0
keras-Preprocessing==1.1.2
matplotlib==3.7.1
numpy==1.22.3
pandas==1.5.3
regex==2023.03.23
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
seaborn==0.12.2
streamlit==1.20.0
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
